pyproject.toml
[project]
name = "dwh-csv-data-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "factory-boy>=3.3.3",
    "faker>=38.2.0",
]
